file directory restructuring
