+++ src/components/pages/NoteResults.js
