+++ src/components/pages/SearchPage.styled.js
