+++ src/components/templates/FeacturedProducts.styled.js
