+++ src/components/templates/Footer.styled.js
