+++ src/components/templates/ProductList.js
@@ -1,8 +1,8 @@
 /* eslint-disable no-unused-vars */
 import { useEffect, useMemo, useState } from "react";
-import { useProductCategories } from "../../../../hooks/useProductCategories";
-import { getProductsByCategoryId } from "../../../../selectors/getProductsByCategory";
-import { Loader } from "../../../../utils/modules/loader/Loader";
+import { useProductCategories } from "../../hooks/useProductCategories";
+import { getProductsByCategoryId } from "../../selectors/getProductsByCategory";
+import { Loader } from "../../utils/modules/loader/Loader";
 import { ProductListContainer } from "./ProductList.styled";
 import { ProductListGrid } from "./ProductListGrid";
 import { ProductSideBar } from "./ProductSideBar";
@@ -15,7 +15,7 @@ export const ProductList = () => {
   const [products, setProducts] = useState([]);
   const [loading, setLoading] = useState(false);
   const [currentPage, setCurrentPage] = useState(1);
-  console.log(products);
+
   useEffect(() => {
     setLoading(true);
     setTimeout(() => {
